Create program to update hiscore from random integer between two numbers to txt file
This program updates hiscore to text. The program generates random integer between two given values and use that integer as score. Whenever the value is hihger than the already generated values, it gets updated in the text file. 

Disclaimer:
This problem is from practice set on YT video from CodewithHarry.